Moved Vocabulary class to Transformer_Train
